[Tree widget]: review public and beta tags on `next`
